chore: rpc -> api modeling plumbing for ib_partition/fabric_monitor
This is similar to #596, #598, #602, and #606, where our shiny new `STYLE_GUIDE.md` expresses using the API layer as the translation boundary between the gRPC and DB layers. As such, gRPC messages should be translated into a model from `api-model` with corresponding `From`/`TryFrom` implementations, and then passed downstream.

This is something we've basically always done, but there are a few cases where things have leaked over time, so I'm just cleaning it up.

This does it for `ib_partition`.
